#1 main.tsをリファクタリングし、ファイル分割を行った。
